Explicitly handle false in makeDbCall()
Fixes #129 
The code's intent was to handle database errors, which return false, but because the code checked for !$outputXml, it interpreted an empty strings as an error, reporting "errors" that were merely empty output. This occurred, for example, when doing a disconnect() with the PDO transport. Now we check for false explicitly.