Make supposedly optional arguments before required args required
This isn't allowed by PHP anymore, and these seemingly were never
called omitting the optionals to the left of required. Changing
the order of the args would be breaking compat and setting a
default value would be semantically inappropriate on some.

The similator funcs/ctors that *did* have their required/optional
arguments in order weren't changed. This should go under some
review to make sure it doesn't break user code in avoidable ways,
but this is required going forward for PHP 8.