i5_connect(): default host to 'localhost'
Adopt a more reliable way to set 'localhost' as the default host if empty $host was specified in i5_connect(). Although the $host parameter of i5_connect() defaulted to 'localhost' in the argument list, this was not a reliable way to ensure 'localhost' for an empty host value. We saw this at a client who used ('', '', '', array(i5_libl...)) and got an error about an unknown host value.