Clean up and add dependencies
This covers what's required and what makes sense to add if you want
a specific transport.

The extensions in requires are ones that might not be always loaded
in every install - I excluded Core, SPL, date, session, ctype, SPL,
and standard since these tend to always be there.

Other changes like to authors aren't made yet. The PHP major version
also isn't quite accurate, but it'll take some time to determine the
exact version, or the courage to bump it next minor release.